KA-137 - Rename ListingOrSpinner files
