+++ onboarding-task/src/components/Loader.tsx
