+++ onboarding-task/src/containers/ListWithLoader.ts
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { IAppState } from '../reducers/IAppState';
-import { IListingOrSpinnerCallbackProps, IListingOrSpinnerDataProps } from '../components/ListingOrSpinner';
-import { ListingOrSpinner as ListingOrSpinnerComponent } from '../components/ListingOrSpinner';
+import { IListingOrSpinnerCallbackProps, IListingOrSpinnerDataProps } from '../components/Loader';
+import { ListingOrSpinner as ListingOrSpinnerComponent } from '../components/Loader';
 import { fetchItems } from '../actionCreators/actionCreators';
 
 const mapStateToProps = (state: IAppState, _ownProps: IListingOrSpinnerDataProps): IListingOrSpinnerDataProps => ({
